[GTNAP-660] feat: Moved HBNAP code to hbnap.c/.h
