+++ contrib/gtnap/gtnap.hbp
@@ -50,5 +50,6 @@ src/gtnap/nap_textview.c
 src/gtnap/nap_window.c
 
 # New HBNAP Api
+src/gtnap/hbnap.c
 src/gtnap/hbnap_forms.c
 src/gtnap/hbnap_menu.c
