+++ contrib/gtnap/src/gtnap/gtnap.c
@@ -75,12 +75,6 @@ typedef struct _gtnap_area_t GtNapArea;
 typedef struct _gtnap_object_t GtNapObject;
 typedef struct _gtnap_geom_t GtNapGeom;
 typedef struct _gtnap_window_t GtNapWindow;
-typedef struct _gtnap_bind_t GtNapBind;
-typedef struct _gtnap_fcolumn_t GtNapFColumn;
-typedef struct _gtnap_fbdconn_t GtNapFDBConn;
-typedef struct _gtnap_farea_t GtNapFArea;
-typedef struct _gtnap_farea2_t GtNapFArea2;
-typedef struct _gtnap_fnode_t GtNapFNode;
 typedef struct _gtnap_prop_t GtNapProp;
 typedef struct _gtnap_t GtNap;
 
@@ -106,8 +100,6 @@ typedef enum _datatype_t
 struct _gtnap_callback_t
 {
     GtNapWindow *gtwin;
-    GtNapForm *form;
-    GtNapMenuItem *menuitem;
     HB_ITEM *block;
     int32_t key;
     uint32_t autoclose_id;
@@ -216,77 +208,6 @@ struct _gtnap_window_t
     ArrPt(GtNapCallback) *callbacks;
 };
 
-struct _gtnap_modal_t
-{
-    uint64_t timestamp;
-    uint32_t close_seconds;
-    GtNapWindow *gtwin;
-};
-
-struct _gtnap_bind_t
-{
-    String *gui_id;
-    PHB_ITEM value;
-    Listener *listener;
-};
-
-struct _gtnap_fcolumn_t
-{
-    align_t align;
-    HB_ITEM *block;
-};
-
-struct _gtnap_fnode_t
-{
-    GtNapFArea2 *area;
-    HB_ULONG recno;
-    bool_t expanded;
-    uint32_t nchildren; /* Child count forward. Stops recursion for collapsed nodes. */
-};
-
-struct _gtnap_fbdconn_t
-{
-    GtNapForm *form;
-    String *cellname;
-    TableView *table;
-    ArrSt(GtNapFArea2) *areas;
-    TreeSt(GtNapFNode) *tdata;
-};
-
-struct _gtnap_farea2_t
-{
-    AREA *area;
-    ArrSt(GtNapFColumn) *columns;
-    HB_ITEM *relkey;
-    HB_USHORT fexpanded;
-};
-
-struct _gtnap_farea_t
-{
-    GtNapForm *form;
-    String *cellname;
-    TableView *table;
-    AREA *area;
-    HB_ULONG cache_recno;     /* Store the DB recno while table drawing */
-    ArrSt(uint32_t) *records; /* Records visible in table (index, deleted, filters) */
-    ArrSt(GtNapFColumn) *columns;
-};
-
-struct _gtnap_form_t
-{
-    NForm *form;
-    Window *window;
-    String *nameid;
-    String *respath;
-    uint32_t modal_ret;
-    GtNapFArea *area;
-    GtNapFDBConn *dbconn;
-    HB_ITEM *OnClose_block;
-    bool_t is_resizable;
-    ArrSt(GtNapBind) *binds;
-    ArrPt(GtNapCallback) *callbacks;
-};
-
 struct _gtnap_prop_t
 {
     String *key;
@@ -313,7 +234,6 @@ struct _gtnap_t
     uint32_t modal_delay_seconds;
     GtNapWindow *modal_time_window;
     ArrPt(GtNapWindow) *windows;
-    ArrPt(GtNapCallback) *menu_callbacks;
     String *debugger_path;
     bool_t debugger_visible;
     GtNapDebugger *debugger;
@@ -324,15 +244,11 @@ struct _gtnap_t
 
 DeclPt(GtNapCallback);
 DeclSt(GtNapColumn);
-DeclSt(GtNapFColumn);
 DeclPt(GtNapArea);
 DeclPt(GtNapObject);
-DeclSt(GtNapBind);
 DeclPt(GtNapWindow);
 DeclPt(GuiComponent);
 DeclSt(GtNapProp);
-DeclSt(GtNapFArea2);
-DeclSt(GtNapFNode);
 
 /*---------------------------------------------------------------------------*/
 
@@ -793,9 +709,7 @@ static void i_gtnap_destroy(GtNap **gtnap)
     cassert_no_null(gtnap);
     cassert_no_null(*gtnap);
     cassert(*gtnap == GTNAP_GLOBAL);
-    cassert(arrpt_size((*gtnap)->menu_callbacks, GtNapCallback) == 0);
     arrpt_destroy(&(*gtnap)->windows, i_destroy_gtwin, GtNapWindow);
-    arrpt_destroy(&(*gtnap)->menu_callbacks, i_destroy_callback, GtNapCallback);
     font_destroy(&(*gtnap)->global_font);
     font_destroy(&(*gtnap)->button_font);
     font_destroy(&(*gtnap)->edit_font);
@@ -1410,7 +1324,6 @@ static GtNap *i_gtnap_create(void)
     GTNAP_GLOBAL->rows = INIT_ROWS;
     GTNAP_GLOBAL->cols = INIT_COLS;
     GTNAP_GLOBAL->windows = arrpt_create(GtNapWindow);
-    GTNAP_GLOBAL->menu_callbacks = arrpt_create(GtNapCallback);
     GTNAP_GLOBAL->date_digits = (hb_setGetCentury() == (HB_BOOL)HB_TRUE) ? 8 : 6;
     GTNAP_GLOBAL->date_chars = GTNAP_GLOBAL->date_digits + 2;
 
@@ -5081,2185 +4994,6 @@ void hb_gtnap_cualib_default_button(const uint32_t nDefault)
     cassert(gtwin->default_button == UINT32_MAX);
     cassert(nDefault > 0);
     gtwin->default_button = nDefault - 1;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static GtNap *i_gtnap_forms_create(void)
-{
-    const char_t *build_cfg = NULL;
-    GTNAP_GLOBAL = heap_new0(GtNap);
-    GTNAP_GLOBAL->date_digits = (hb_setGetCentury() == (HB_BOOL)HB_TRUE) ? 8 : 6;
-    GTNAP_GLOBAL->date_chars = GTNAP_GLOBAL->date_digits + 2;
-
-    {
-        char_t path[512];
-        bfile_dir_work(path, sizeof(path));
-        GTNAP_GLOBAL->working_path = str_c(path);
-    }
-
-#if defined(__DEBUG__)
-    build_cfg = "Debug";
-#else
-    build_cfg = "Release";
-#endif
-
-    {
-        const char_t *debpath = deblib_path();
-#if defined(__MACOS__)
-        GTNAP_GLOBAL->debugger_path = str_cpath("%s/%s/bin/gtnapdeb.app/Contents/MacOS/gtnapdeb", debpath, build_cfg);
-#else
-        GTNAP_GLOBAL->debugger_path = str_cpath("%s/%s/bin/gtnapdeb", debpath, build_cfg);
-#endif
-        GTNAP_GLOBAL->debugger_visible = FALSE;
-        GTNAP_GLOBAL->debugger = NULL;
-    }
-
-    {
-        PHB_ITEM ritem = NULL;
-        deblib_init_colors(i_COLORS);
-        ritem = hb_itemDo(INIT_CODEBLOCK, 0);
-        hb_itemRelease(ritem);
-    }
-
-    hb_itemRelease(INIT_CODEBLOCK);
-    INIT_TITLE[0] = 0;
-    INIT_CODEBLOCK = NULL;
-    return GTNAP_GLOBAL;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_gtnap_forms_update(GtNap *gtnap, const real64_t prtime, const real64_t ctime)
-{
-    cassert(gtnap == NULL || gtnap == GTNAP_GLOBAL);
-    gtnap = GTNAP_GLOBAL;
-    cassert_no_null(gtnap);
-    unref(prtime);
-    unref(ctime);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_gtnap_forms_destroy(GtNap **gtnap)
-{
-    cassert_no_null(gtnap);
-    cassert_no_null(*gtnap);
-    cassert(*gtnap == GTNAP_GLOBAL);
-    str_destroy(&(*gtnap)->working_path);
-    str_destroy(&(*gtnap)->debugger_path);
-
-    if ((*gtnap)->debugger != NULL)
-        nap_debugger_destroy(&(*gtnap)->debugger);
-
-    nforms_finish();
-    heap_delete(&(*gtnap), GtNap);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_init_app(HB_ITEM *main_block)
-{
-    void *hInstance = NULL;
-
-#if defined(HB_OS_WIN)
-    hb_winmainArgGet(&hInstance, NULL, NULL);
-#endif
-
-    nforms_start();
-    INIT_CODEBLOCK = hb_itemNew(main_block);
-
-    osmain_imp(
-        0, NULL, hInstance, 0.5f,
-        (FPtr_app_create)i_gtnap_forms_create,
-        (FPtr_app_update)i_gtnap_forms_update,
-        (FPtr_destroy)i_gtnap_forms_destroy,
-        (char_t *)"");
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_exit_app(void)
-{
-    osapp_finish();
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnFormMoved(GtNapForm *form, Event *e)
-{
-    const EvPos *p = event_params(e, EvPos);
-    cassert_no_null(form);
-    i_write_prop_r32(tc(form->nameid), i_XPOS_PROP, p->x);
-    i_write_prop_r32(tc(form->nameid), i_YPOS_PROP, p->y);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnFormResize(GtNapForm *form, Event *e)
-{
-    const EvSize *p = event_params(e, EvSize);
-    Window *window = event_sender(e, Window);
-    cassert_no_null(form);
-    if (window_get_maximize(window) == TRUE)
-    {
-        i_write_prop_r32(tc(form->nameid), i_WIDTH_PROP, i_MAXIMIZED_SIZE);
-        i_write_prop_r32(tc(form->nameid), i_HEIGHT_PROP, i_MAXIMIZED_SIZE);
-    }
-    else
-    {
-        i_write_prop_r32(tc(form->nameid), i_WIDTH_PROP, p->width);
-        i_write_prop_r32(tc(form->nameid), i_HEIGHT_PROP, p->height);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-GtNapForm *hbnap_forms_load(const char_t *pathname, const char_t *resource_path, const uint32_t flags)
-{
-    NForm *form = nform_from_file(pathname, NULL);
-    if (form != NULL)
-    {
-        GtNapForm *gtform = heap_new0(GtNapForm);
-        str_split_pathext(pathname, NULL, &gtform->nameid, NULL);
-        gtform->form = form;
-        gtform->respath = str_c(resource_path);
-        gtform->binds = arrst_create(GtNapBind);
-        gtform->callbacks = arrpt_create(GtNapCallback);
-
-        if (!(flags & HBNAP_FORMS_EMBEDDED_PANEL))
-        {
-            uint32_t nflags = 0;
-            if (flags & HBNAP_FORMS_RESIZABLE)
-            {
-                nflags |= ekWINDOW_STDRES;
-                gtform->is_resizable = TRUE;
-            }
-            else
-            {
-                nflags |= ekWINDOW_STD;
-            }
-
-            if (flags & HBNAP_FORMS_CLOSE_ON_ESC)
-                nflags |= ekWINDOW_ESC;
-
-            if (flags & HBNAP_FORMS_CLOSE_ON_RETURN)
-                nflags |= ekWINDOW_RETURN;
-
-            gtform->window = nform_window(gtform->form, nflags, resource_path);
-            if (gtform->window != NULL)
-            {
-                window_OnMoved(gtform->window, listener(gtform, i_OnFormMoved, GtNapForm));
-                window_OnResize(gtform->window, listener(gtform, i_OnFormResize, GtNapForm));
-            }
-            else
-            {
-                hbnap_forms_destroy(&gtform);
-            }
-        }
-
-        return gtform;
-    }
-
-    return NULL;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_remove_fcolumn(GtNapFColumn *column)
-{
-    cassert_no_null(column);
-    if (column->block != NULL)
-        hb_itemRelease(column->block);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_destroy_farea(GtNapFArea **area)
-{
-    cassert_no_null(area);
-    (*area)->form = NULL;
-    (*area)->table = NULL;
-    (*area)->area = NULL;
-    str_destroy(&(*area)->cellname);
-    arrst_destroy(&(*area)->records, NULL, uint32_t);
-    arrst_destroy(&(*area)->columns, i_remove_fcolumn, GtNapFColumn);
-    heap_delete(area, GtNapFArea);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_remove_fnode(GtNapFNode *node)
-{
-    cassert_no_null(node);
-    node->area = NULL;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_remove_farea2(GtNapFArea2 *area)
-{
-    cassert_no_null(area);
-    area->area = NULL; /* The life cycle of the area is managed exclusively by Harbour */
-
-    arrst_destroy(&area->columns, i_remove_fcolumn, GtNapFColumn);
-
-    if (area->relkey != NULL)
-        hb_itemRelease(area->relkey);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_destroy_fdbconn(GtNapFDBConn **dbconn)
-{
-    cassert_no_null(dbconn);
-    (*dbconn)->form = NULL;
-    (*dbconn)->table = NULL;
-    str_destroy(&(*dbconn)->cellname);
-    treest_destroy(&(*dbconn)->tdata, i_remove_fnode, GtNapFNode);
-    arrst_destroy(&(*dbconn)->areas, i_remove_farea2, GtNapFArea2);
-    heap_delete(dbconn, GtNapFDBConn);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_remove_bind(GtNapBind *bind)
-{
-    cassert_no_null(bind);
-    str_destroy(&bind->gui_id);
-    if (bind->value != NULL)
-        hb_itemRelease(bind->value);
-    if (bind->listener != NULL)
-        listener_destroy(&bind->listener);
-    bind->value = NULL;
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_destroy(GtNapForm **form)
-{
-    cassert_no_null(form);
-    cassert_no_null(*form);
-    str_destroy(&(*form)->nameid);
-    str_destroy(&(*form)->respath);
-    ptr_destopt(window_destroy, &(*form)->window, Window);
-
-    if ((*form)->OnClose_block != NULL)
-    {
-        hb_itemRelease((*form)->OnClose_block);
-        (*form)->OnClose_block = NULL;
-    }
-
-    arrst_destroy(&(*form)->binds, i_remove_bind, GtNapBind);
-    arrpt_destroy(&(*form)->callbacks, i_destroy_callback, GtNapCallback);
-    ptr_destopt(i_destroy_farea, &(*form)->area, GtNapFArea);
-    ptr_destopt(i_destroy_fdbconn, &(*form)->dbconn, GtNapFDBConn);
-    nform_destroy(&(*form)->form);
-    heap_delete(form, GtNapForm);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_title(GtNapForm *form, const char_t *text)
-{
-    cassert_no_null(form);
-    window_title(form->window, text);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_set_text(GtNapForm *form, const char_t *cell, const char_t *text)
-{
-    cassert_no_null(form);
-    nform_set_control_str(form->form, cell, text);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_set_int(GtNapForm *form, const char_t *cell, const int32_t value)
-{
-    cassert_no_null(form);
-    nform_set_control_int(form->form, cell, value);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_insert_text(GtNapForm *form, const char_t *cell, const char_t *text)
-{
-    cassert_no_null(form);
-    nform_add_control_str(form->form, cell, text);
-}
-
-/*---------------------------------------------------------------------------*/
-
-int32_t hbnap_forms_get_int(GtNapForm *form, const char_t *cell)
-{
-    int32_t value = 0;
-    cassert_no_null(form);
-    if (nform_get_control_int(form->form, cell, &value) == TRUE)
-        return value;
-    return -1;
-}
-
-/*---------------------------------------------------------------------------*/
-
-bool_t hbnap_forms_embed(GtNapForm *form, GtNapForm *embedded_form, const char_t *cell)
-{
-    bool_t ok = FALSE;
-    cassert_no_null(form);
-    cassert_no_null(embedded_form);
-    cassert_no_null(form->window);
-    cassert(embedded_form->window == NULL);
-    ok = nform_change_embedded_panel(form->form, embedded_form->form, tc(embedded_form->respath), cell);
-    if (ok == TRUE)
-        window_update(form->window);
-    return ok;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static const char_t *i_farea_eval_field(GtNapFArea *area, const uint32_t field_id, const uint32_t row_id)
-{
-    uint32_t recno = 0;
-    const GtNapFColumn *column = NULL;
-    HB_ITEM *ritem = NULL;
-
-    cassert_no_null(area);
-    cassert(field_id > 0);
-
-    /* Go to DB record */
-    recno = *arrst_get_const(area->records, row_id, uint32_t);
-    SELF_GOTO(area->area, recno);
-
-    /* Get the table column */
-    column = arrst_get_const(area->columns, field_id - 1, GtNapFColumn);
-
-    /* CodeBlock that computes the cell content */
-    ritem = hb_itemDo(column->block, 0);
-
-    /* Fill the temporal cell buffer with cell result */
-    hb_item_to_char(ritem, TEMP_BUFFER, sizeof(TEMP_BUFFER), TRUE);
-
-    hb_itemRelease(ritem);
-    return TEMP_BUFFER;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnTableFAreaData(GtNapFArea *area, Event *e)
-{
-    uint32_t etype = event_type(e);
-    cassert_no_null(area);
-
-    switch (etype)
-    {
-    case ekGUI_EVENT_TBL_BEGIN:
-        SELF_RECNO(area->area, &area->cache_recno);
-        break;
-
-    case ekGUI_EVENT_TBL_END:
-        SELF_GOTO(area->area, area->cache_recno);
-        area->cache_recno = UINT32_MAX;
-        break;
-
-    case ekGUI_EVENT_TBL_NROWS:
-    {
-        uint32_t *n = event_result(e, uint32_t);
-        *n = arrst_size(area->records, uint32_t);
-        break;
-    }
-
-    case ekGUI_EVENT_TBL_CELL:
-    {
-        EvTbCell *cell = event_result(e, EvTbCell);
-        const EvTbPos *pos = event_params(e, EvTbPos);
-        cell->text = i_farea_eval_field(area, pos->col + 1, pos->row);
-        break;
-    }
-
-    default:
-        break;
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_farea_refresh(GtNapFArea *area)
-{
-    HB_ULONG ulCurRec;
-
-    cassert_no_null(area);
-
-    /* Database current RECNO() */
-    SELF_RECNO(area->area, &ulCurRec);
-
-    /* Clear the current record index */
-    arrst_clear(area->records, NULL, uint32_t);
-
-    /* Generate the record index for TableView */
-    // if (area->while_block == NULL)
-    {
-        HB_BOOL fEof;
-        SELF_GOTOP(area->area);
-        SELF_EOF(area->area, &fEof);
-        while (fEof == HB_FALSE)
-        {
-            HB_ULONG uiRecNo = 0;
-            SELF_RECNO(area->area, &uiRecNo);
-            arrst_append(area->records, (uint32_t)uiRecNo, uint32_t);
-            SELF_SKIP(area->area, 1);
-            SELF_EOF(area->area, &fEof);
-        }
-    }
-    // else
-    //{
-    //     HB_BOOL fEof;
-    //     SELF_GOTOP(area->area);
-    //     SELF_EOF(area->area, &fEof);
-    //     while (fEof == HB_FALSE)
-    //     {
-    //         HB_ULONG uiRecNo = 0;
-    //         SELF_RECNO(area->area, &uiRecNo);
-
-    //        {
-    //            PHB_ITEM ritem = hb_itemDo(area->while_block, 0);
-    //            HB_TYPE type = HB_ITEM_TYPE(ritem);
-    //            bool_t add = FALSE;
-    //            cassert_unref(type == HB_IT_LOGICAL, type);
-    //            add = (bool_t)hb_itemGetL(ritem);
-    //            hb_itemRelease(ritem);
-
-    //            if (add == TRUE)
-    //                arrst_append(area->records, (uint32_t)uiRecNo, uint32_t);
-    //        }
-
-    //        SELF_SKIP(area->area, 1);
-    //        SELF_EOF(area->area, &fEof);
-    //    }
-    //}
-
-    /* Restore database RECNO() */
-    SELF_GOTO(area->area, ulCurRec);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static uint32_t i_frow_from_recno(GtNapFArea *area, const uint32_t recno)
-{
-    cassert_no_null(area);
-    cassert(recno > 0);
-    arrst_foreach_const(rec, area->records, uint32_t)
-        if (*rec == recno)
-            return rec_i;
-    arrst_end();
-    return UINT32_MAX;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_farea_select_row(GtNapFArea *area)
-{
-    HB_ULONG ulCurRec;
-    uint32_t sel_row;
-    // TableView *view;
-
-    cassert_no_null(area);
-    // cassert_no_null(gtarea->gtobj);
-    // cassert(gtarea->gtobj->type == ekOBJ_TABLEVIEW);
-    // view = (TableView *)gtarea->gtobj->component;
-
-    /* Current selected */
-    SELF_RECNO(area->area, &ulCurRec);
-
-    sel_row = i_frow_from_recno(area, (uint32_t)ulCurRec);
-
-    /* In multisel table, the selected rows comes from  VN_Selecio */
-    // if (gtarea->gtobj->multisel == TRUE)
-    //{
-    //     if (tableview_get_focus_row(view) == UINT32_MAX)
-    //     {
-    //         /* We use RECNO for focused row */
-    //         if (sel_row != UINT32_MAX)
-    //         {
-    //             tableview_focus_row(view, sel_row, ekTOP);
-    //         }
-    //         else
-    //         {
-    //             uint32_t nrecs = arrst_size(gtarea->records, uint32_t);
-    //             sel_row = tableview_get_focus_row(view);
-    //             /* We move recno to current focused row */
-    //             if (sel_row >= nrecs)
-    //             {
-    //                 sel_row = 0;
-    //             }
-
-    //            if (sel_row < nrecs)
-    //            {
-    //                uint32_t recno = *arrst_get_const(gtarea->records, sel_row, uint32_t);
-    //                tableview_select(view, &sel_row, 1);
-    //                SELF_GOTO(gtarea->area, recno);
-    //            }
-    //        }
-    //    }
-    //}
-    // else
-    {
-        tableview_deselect_all(area->table);
-
-        if (sel_row != UINT32_MAX)
-        {
-            tableview_select(area->table, &sel_row, 1);
-            tableview_focus_row(area->table, sel_row, ekTOP);
-        }
-        /* RECNO() doesn't exists in view (perhaps is deleted) */
-        else
-        {
-            uint32_t nrecs = arrst_size(area->records, uint32_t);
-            sel_row = tableview_get_focus_row(area->table);
-            /* We move recno to current focused row */
-            if (sel_row < nrecs)
-            {
-                uint32_t recno = *arrst_get_const(area->records, sel_row, uint32_t);
-                tableview_select(area->table, &sel_row, 1);
-                SELF_GOTO(area->area, recno);
-            }
-        }
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_map_bind_area_to_form(GtNapFArea *area)
-{
-    cassert_no_null(area);
-    cassert(area->table == NULL);
-    cassert_no_null(area->form);
-    area->table = nform_get_tableview(area->form->form, tc(area->cellname));
-    if (area->table != NULL)
-    {
-        tableview_OnData(area->table, listener(area, i_OnTableFAreaData, GtNapFArea));
-        i_farea_refresh(area);
-        tableview_update(area->table);
-        i_farea_select_row(area);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static GtNapFArea *i_create_farea(GtNapForm *form, AREA *area)
-{
-    GtNapFArea *farea = heap_new0(GtNapFArea);
-    farea->form = form;
-    farea->cellname = NULL;
-    farea->table = NULL;
-    farea->area = area;
-    farea->cache_recno = UINT32_MAX;
-    farea->records = arrst_create(uint32_t);
-    farea->columns = arrst_create(GtNapFColumn);
-    return farea;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static GtNapFDBConn *i_create_dbconn(GtNapForm *form, const char_t *cell)
-{
-    GtNapFDBConn *dbconn = heap_new0(GtNapFDBConn);
-    dbconn->form = form;
-    dbconn->cellname = str_c(cell);
-    dbconn->table = NULL;
-    dbconn->tdata = treest_create(GtNapFNode);
-    dbconn->areas = arrst_create(GtNapFArea2);
-    return dbconn;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_map_bind_to_form(NForm *form, ArrSt(GtNapBind) *binds)
-{
-    arrst_foreach(bind, binds, GtNapBind)
-        if (bind->value != NULL)
-        {
-            PHB_ITEM base = bind->value;
-
-            if (HB_ITEM_TYPE(base) == HB_IT_BYREF)
-                base = hb_itemUnRef(base);
-
-            if (HB_ITEM_TYPE(base) == HB_IT_STRING)
-            {
-                const char_t *text = hb_itemGetCPtr(base);
-                String *str = str_c(text);
-                i_rtrim(tcc(str));
-                nform_set_control_str(form, tc(bind->gui_id), tc(str));
-                str_destroy(&str);
-            }
-            else if (HB_ITEM_TYPE(base) == HB_IT_LOGICAL)
-            {
-                HB_BOOL value = hb_itemGetL(base);
-                nform_set_control_bool(form, tc(bind->gui_id), (bool_t)value);
-            }
-            else if (HB_ITEM_TYPE(base) == HB_IT_INTEGER)
-            {
-                long value = hb_itemGetNL(base);
-                nform_set_control_int(form, tc(bind->gui_id), (int32_t)value);
-            }
-            else if (HB_ITEM_TYPE(base) == HB_IT_DOUBLE)
-            {
-                double value = hb_itemGetND(base);
-                nform_set_control_real(form, tc(bind->gui_id), (real32_t)value);
-            }
-        }
-        else if (bind->listener != NULL)
-        {
-            if (nform_set_listener(form, tc(bind->gui_id), bind->listener) == TRUE)
-                bind->listener = NULL;
-        }
-    arrst_end()
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_bind(GtNapForm *form, HB_ITEM *cell_bind)
-{
-    HB_SIZE i, n = UINT32_MAX;
-    cassert_no_null(form);
-    cassert(HB_ITEM_TYPE(cell_bind) == HB_IT_ARRAY);
-    n = hb_arrayLen(cell_bind);
-    for (i = 1; i <= n; ++i)
-    {
-        PHB_ITEM bind_item = hb_arrayGetItemPtr(cell_bind, i);
-        PHB_ITEM name_item = NULL;
-        PHB_ITEM var_item = NULL;
-        const char *gui_id = NULL;
-        GtNapBind *bind = arrst_new0(form->binds, GtNapBind);
-        cassert(HB_ITEM_TYPE(bind_item) == HB_IT_ARRAY);
-        cassert(hb_arrayLen(bind_item) == 2);
-        name_item = hb_arrayGetItemPtr(bind_item, 1);
-        var_item = hb_arrayGetItemPtr(bind_item, 2);
-        cassert(HB_ITEM_TYPE(name_item) == HB_IT_STRING);
-        gui_id = hb_itemGetCPtr(name_item);
-        bind->gui_id = str_c(cast_const(gui_id, char_t));
-        bind->value = hb_itemNew(var_item);
-    }
-
-    i_map_bind_to_form(form->form, form->binds);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_bind_store(GtNapForm *form)
-{
-    cassert_no_null(form);
-    arrst_foreach(bind, form->binds, GtNapBind)
-        if (bind->value != NULL)
-        {
-            /* We only can save a value in HB_IT_BYREF types */
-            if (HB_ITEM_TYPE(bind->value) == HB_IT_BYREF)
-            {
-                PHB_ITEM base = hb_itemUnRef(bind->value);
-                if (HB_ITEM_TYPE(base) == HB_IT_STRING)
-                {
-                    const char_t *text = NULL;
-                    if (nform_get_control_str(form->form, tc(bind->gui_id), &text) == TRUE)
-                        hb_itemPutC(base, text);
-                }
-                else if (HB_ITEM_TYPE(base) == HB_IT_LOGICAL)
-                {
-                    bool_t value = FALSE;
-                    if (nform_get_control_bool(form->form, tc(bind->gui_id), &value) == TRUE)
-                        hb_itemPutL(base, value ? HB_TRUE : HB_FALSE);
-                }
-                else if (HB_ITEM_TYPE(base) == HB_IT_INTEGER)
-                {
-                    int32_t value = 0;
-                    if (nform_get_control_int(form->form, tc(bind->gui_id), &value) == TRUE)
-                        hb_itemPutNL(base, (long)value);
-                }
-                else if (HB_ITEM_TYPE(base) == HB_IT_DOUBLE)
-                {
-                    real32_t value = 0;
-                    if (nform_get_control_real(form->form, tc(bind->gui_id), &value) == TRUE)
-                        hb_itemPutND(base, (double)value);
-                }
-            }
-        }
-    arrst_end()
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_area_bind(GtNapForm *form, HB_ITEM *column_bind)
-{
-    AREA *area = NULL;
-    HB_SIZE n = UINT32_MAX;
-    cassert_no_null(form);
-    cassert(form->area == NULL);
-    cassert(HB_ITEM_TYPE(column_bind) == HB_IT_ARRAY);
-    n = hb_arrayLen(column_bind);
-    cassert(n > 1);
-
-    area = cast(hb_rddGetCurrentWorkAreaPointer(), AREA);
-    if (area != NULL)
-    {
-        HB_SIZE i;
-        form->area = i_create_farea(form, area);
-
-        for (i = 2; i <= n; ++i)
-        {
-            PHB_ITEM bind_item = hb_arrayGetItemPtr(column_bind, i);
-            PHB_ITEM block_item = NULL;
-            GtNapFColumn *column = NULL;
-            /* At the moment, the column-bind item only has one member. The column block */
-            cassert(HB_ITEM_TYPE(bind_item) == HB_IT_ARRAY);
-            cassert(hb_arrayLen(bind_item) == 1);
-            block_item = hb_arrayGetItemPtr(bind_item, 1);
-            cassert(HB_ITEM_TYPE(block_item) == HB_IT_BLOCK);
-            column = arrst_new0(form->area->columns, GtNapFColumn);
-            column->block = block_item ? hb_itemNew(block_item) : NULL;
-        }
-
-        /* The first element in bind array is the tableView cell name */
-        {
-            PHB_ITEM name_item = hb_arrayGetItemPtr(column_bind, 1);
-            const char *gui_id = NULL;
-            cassert(HB_ITEM_TYPE(name_item) == HB_IT_STRING);
-            gui_id = hb_itemGetCPtr(name_item);
-            form->area->cellname = str_c(cast_const(gui_id, char_t));
-            i_map_bind_area_to_form(form->area);
-        }
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_area_refresh(GtNapForm *form)
-{
-    cassert_no_null(form);
-    cassert_no_null(form->area);
-    i_farea_refresh(form->area);
-    tableview_update(form->area->table);
-    i_farea_select_row(form->area);
-}
-
-/*---------------------------------------------------------------------------*/
-
-uint32_t hbnap_forms_area_recno(GtNapForm *form)
-{
-    const ArrSt(uint32_t) *sel = 0;
-    cassert_no_null(form);
-    cassert_no_null(form->area);
-    sel = tableview_selected(form->area->table);
-    if (arrst_size(sel, uint32_t) == 1)
-    {
-        uint32_t index = *arrst_first_const(sel, uint32_t);
-        return *arrst_get(form->area->records, index, uint32_t);
-    }
-    else
-    {
-        return UINT32_MAX;
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static align_t i_hbalign(int hbalign)
-{
-    switch (hbalign)
-    {
-    case HBNAP_LEFT:
-        return ekLEFT;
-    case HBNAP_CENTER:
-        return ekCENTER;
-    case HBNAP_RIGHT:
-        return ekRIGHT;
-    default:
-        cassert_default(hbalign);
-    }
-
-    return ekLEFT;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static bool_t i_items_equal(PHB_ITEM a, PHB_ITEM b)
-{
-    HB_TYPE ta = HB_ITEM_TYPE(a);
-    HB_TYPE tb = HB_ITEM_TYPE(b);
-
-    /* Numeric subtypes (INTEGER, LONG, DOUBLE) are interchangeable:
-     * DBOI_KEYVAL always returns DOUBLE while code blocks may return INTEGER */
-    if ((ta & HB_IT_NUMERIC) && (tb & HB_IT_NUMERIC))
-        return (bool_t)(hb_itemGetND(a) == hb_itemGetND(b));
-
-    cassert(ta == tb);
-
-    if (ta & HB_IT_STRING)
-        return (bool_t)(hb_stricmp(hb_itemGetCPtr(a), hb_itemGetCPtr(b)) == 0);
-    if (ta & HB_IT_LOGICAL)
-        return (bool_t)(hb_itemGetL(a) == hb_itemGetL(b));
-
-    cassert(FALSE);
-    return FALSE;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static uint32_t i_count_db_children(GtNapFArea2 *parea, HB_ULONG precno, GtNapFArea2 *carea)
-{
-    PHB_ITEM key = NULL;
-    HB_BOOL found = HB_FALSE;
-    uint32_t count = 0;
-    SELF_GOTO(parea->area, precno);
-    key = hb_itemDo(parea->relkey, 0);
-    SELF_SEEK(carea->area, HB_FALSE, key, HB_FALSE);
-    SELF_FOUND(carea->area, &found);
-    if (found == HB_TRUE)
-    {
-        HB_BOOL eof = HB_FALSE;
-        SELF_EOF(carea->area, &eof);
-        while (eof == HB_FALSE)
-        {
-            DBORDERINFO kinfo;
-            bool_t equ = FALSE;
-            memset(&kinfo, 0, sizeof(kinfo));
-            kinfo.itmResult = hb_itemNew(NULL);
-            SELF_ORDINFO(carea->area, DBOI_KEYVAL, &kinfo);
-            equ = i_items_equal(key, kinfo.itmResult);
-            hb_itemRelease(kinfo.itmResult);
-            if (equ == TRUE)
-            {
-                SELF_SKIP(carea->area, 1);
-                SELF_EOF(carea->area, &eof);
-                count++;
-            }
-            else
-            {
-                break;
-            }
-        }
-    }
-
-    hb_itemRelease(key);
-    return count;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_build_children(GtNapFDBConn *dbconn, NodeSt(GtNapFNode) *parent, uint32_t level)
-{
-    GtNapFArea2 *parea = NULL;
-    GtNapFArea2 *carea = NULL;
-    GtNapFNode *pdata = NULL;
-    PHB_ITEM key = NULL;
-    HB_BOOL found = HB_FALSE;
-
-    cassert_no_null(dbconn);
-    parea = arrst_get(dbconn->areas, level, GtNapFArea2);
-    carea = arrst_get(dbconn->areas, level + 1, GtNapFArea2);
-    pdata = treest_node_data(parent, GtNapFNode);
-
-    /* Position parent at its record so relkey block reads the right value */
-    SELF_GOTO(parea->area, pdata->recno);
-
-    /* Evaluate parent relkey block, seek key for child area */
-    key = hb_itemDo(parea->relkey, 0);
-
-    /*
-    i_dump_area_tags(carea->area, "INVOICES");
-    i_dump_area_order(carea->area, "Label");
-    */
-
-    /*
-     * SEEK in child area (order already set).
-     * Hard seek (HB_FALSE remains in EOF if it not found)
-     * Find last (HB_FALSE find the first record match)
-     */
-    SELF_SEEK(carea->area, HB_FALSE, key, HB_FALSE);
-
-    SELF_FOUND(carea->area, &found);
-    if (found == HB_TRUE)
-    {
-        HB_BOOL eof = HB_FALSE;
-        uint32_t nareas = arrst_size(dbconn->areas, GtNapFArea2);
-        SELF_EOF(carea->area, &eof);
-        while (eof == HB_FALSE)
-        {
-            /* Child key must still match parent key (read from active index) */
-            DBORDERINFO kinfo;
-            bool_t equ = FALSE;
-            memset(&kinfo, 0, sizeof(kinfo));
-            kinfo.itmResult = hb_itemNew(NULL);
-            SELF_ORDINFO(carea->area, DBOI_KEYVAL, &kinfo);
-            equ = i_items_equal(key, kinfo.itmResult);
-            hb_itemRelease(kinfo.itmResult);
-
-            if (equ == TRUE)
-            {
-                /* Add child node to tree */
-                NodeSt(GtNapFNode) *child = treest_node_insert(parent, UINT32_MAX, GtNapFNode);
-                GtNapFNode *cdata = treest_node_data(child, GtNapFNode);
-                HB_ULONG crecno = 0;
-
-                SELF_RECNO(carea->area, &crecno);
-                cdata->area = carea;
-                cdata->recno = crecno;
-                if (carea->fexpanded != 0)
-                {
-                    PHB_ITEM pExp = hb_itemNew(NULL);
-                    SELF_GETVALUE(carea->area, carea->fexpanded, pExp);
-                    cdata->expanded = (bool_t)hb_itemGetL(pExp);
-                    hb_itemRelease(pExp);
-                }
-                else
-                {
-                    cdata->expanded = FALSE;
-                }
-
-                cdata->nchildren = 0;
-                /* Exists grandchildren */
-                if (level + 2 < nareas)
-                {
-                    /* Recurse for grandchildren only if expanded, collapsed nodes get a count without building */
-                    if (cdata->expanded == TRUE)
-                    {
-                        i_build_children(dbconn, child, level + 1);
-                    }
-                    else
-                    {
-                        GtNapFArea2 *gcarea = arrst_get(dbconn->areas, level + 2, GtNapFArea2);
-                        cdata->nchildren = i_count_db_children(carea, crecno, gcarea);
-                    }
-                }
-
-                /* Restore child cursor after recursion, then advance */
-                SELF_GOTO(carea->area, crecno);
-                SELF_SKIP(carea->area, 1);
-                SELF_EOF(carea->area, &eof);
-            }
-            else
-            {
-                break;
-            }
-        }
-    }
-
-    hb_itemRelease(key);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static uint32_t i_area_level(const GtNapFDBConn *dbconn, const GtNapFArea2 *area)
-{
-    cassert_no_null(dbconn);
-    arrst_foreach_const(larea, dbconn->areas, GtNapFArea2)
-        if (larea == area)
-            return larea_i;
-    arrst_end();
-    cassert(FALSE);
-    return UINT32_MAX;
-}
-
-/*---------------------------------------------------------------------------*/
-
-/*
-static void i_dump_node_count(const TreeSt(GtNapFNode) *tree)
-{
-    uint32_t n = 0;
-    treest_foreach_const(node, tree, GtNapFNode)
-        n += 1;
-    treest_end()
-    bstd_printf("Node count: %d\n", n);
-}
-*/
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnTreeFAreaData(GtNapFDBConn *dbconn, Event *e)
-{
-    uint32_t etype = event_type(e);
-    cassert_no_null(dbconn);
-
-    switch (etype)
-    {
-
-    case ekGUI_EVENT_TBL_NROOTS:
-    {
-        uint32_t *nroots = event_result(e, uint32_t);
-        NodeSt(GtNapFNode) *root = treest_root_get(dbconn->tdata, GtNapFNode);
-        *nroots = root ? treest_node_size(root, GtNapFNode) : 0;
-        break;
-    }
-
-    case ekGUI_EVENT_TBL_NODEINFO:
-    {
-        const EvTbNode *node = event_params(e, EvTbNode);
-        EvTbNodeInfo *info = event_result(e, EvTbNodeInfo);
-        NodeSt(GtNapFNode) *parent = cast(node->parent, NodeSt(GtNapFNode));
-        NodeSt(GtNapFNode) *child = NULL;
-        GtNapFNode *data = NULL;
-
-        if (parent == NULL)
-            parent = treest_root_get(dbconn->tdata, GtNapFNode);
-
-        child = treest_node_get(parent, node->child, GtNapFNode);
-        data = treest_node_data(child, GtNapFNode);
-        info->node = child;
-
-        {
-            /* If children are built, use tree count, otherwise use the pre-counted DB children */
-            uint32_t nc = treest_node_size(child, GtNapFNode);
-            cassert((nc > 0 && data->nchildren == 0) || nc == 0);
-            info->nchildren = (nc > 0) ? nc : data->nchildren;
-        }
-
-        info->expanded = data->expanded;
-        break;
-    }
-
-    case ekGUI_EVENT_TBL_EXPAND:
-    {
-        const EvTbExpand *p = event_params(e, EvTbExpand);
-        NodeSt(GtNapFNode) *node = cast(p->node, NodeSt(GtNapFNode));
-        GtNapFNode *data = treest_node_data(node, GtNapFNode);
-        cassert_no_null(data);
-        cassert_no_null(data->area);
-        data->expanded = p->expanded;
-
-        /* Expanding a node whose subtree was counted but not built */
-        if (p->expanded == TRUE && data->nchildren > 0 && treest_node_size(node, GtNapFNode) == 0)
-        {
-            uint32_t level = i_area_level(dbconn, data->area);
-            i_build_children(dbconn, node, level);
-            /*i_dump_node_count(dbconn->tdata);*/
-        }
-
-        /* Mark expanded/collapsed in DB */
-        if (data->area->fexpanded != 0)
-        {
-            DBLOCKINFO linfo;
-            HB_ERRCODE hbres;
-            PHB_ITEM pVal = hb_itemPutL(NULL, (HB_BOOL)p->expanded);
-            SELF_GOTO(data->area->area, data->recno);
-            memset(&linfo, 0, sizeof(linfo));
-            linfo.uiMethod = REC_LOCK;
-            hbres = SELF_LOCK(data->area->area, &linfo);
-            cassert_unref(hbres == HB_SUCCESS, hbres);
-
-            if (linfo.fResult)
-            {
-                hbres = SELF_PUTVALUE(data->area->area, data->area->fexpanded, pVal);
-                cassert_unref(hbres == HB_SUCCESS, hbres);
-                linfo.uiMethod = REC_UNLOCK;
-                hbres = SELF_LOCK(data->area->area, &linfo);
-                cassert_unref(hbres == HB_SUCCESS, hbres);
-            }
-
-            hb_itemRelease(pVal);
-        }
-        break;
-    }
-
-    case ekGUI_EVENT_TBL_CELL:
-    {
-        const EvTbPos *pos = event_params(e, EvTbPos);
-        NodeSt(GtNapFNode) *node = cast(pos->node, NodeSt(GtNapFNode));
-        GtNapFNode *data = treest_node_data(node, GtNapFNode);
-        const GtNapFColumn *col = arrst_get_const(data->area->columns, pos->col, GtNapFColumn);
-
-        /* Position current node cursor */
-        SELF_GOTO(data->area->area, data->recno);
-
-        /* Position all ancestor cursors (bottom-up walk, skip virtual root) */
-        {
-            NodeSt(GtNapFNode) *anode = treest_node_parent(node, GtNapFNode);
-            while (anode != NULL)
-            {
-                GtNapFNode *adata = treest_node_data(anode, GtNapFNode);
-                if (adata->area != NULL)
-                    SELF_GOTO(adata->area->area, adata->recno);
-                anode = treest_node_parent(anode, GtNapFNode);
-            }
-        }
-
-        {
-            EvTbCell *cell = event_result(e, EvTbCell);
-            PHB_ITEM ritem = hb_itemDo(col->block, 0);
-            hb_item_to_char(ritem, TEMP_BUFFER, sizeof(TEMP_BUFFER), TRUE);
-            hb_itemRelease(ritem);
-            cell->text = TEMP_BUFFER;
-            cell->align = col->align;
-        }
-
-        break;
-    }
-
-    default:
-        break;
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static HB_USHORT i_area_expanded_field(AREA *area)
-{
-    HB_USHORT n = 0;
-    HB_USHORT i;
-    char szName[64];
-    SELF_FIELDCOUNT(area, &n);
-    for (i = 1; i <= n; ++i)
-    {
-        SELF_FIELDNAME(area, i, szName);
-        if (hb_stricmp(szName, "EXPANDED") == 0)
-        {
-            PHB_ITEM pType = hb_itemNew(NULL);
-            bool_t islogic = FALSE;
-            SELF_FIELDINFO(area, i, DBS_TYPE, pType);
-            islogic = (bool_t)(hb_itemGetCPtr(pType)[0] == 'L');
-            hb_itemRelease(pType);
-            return islogic ? i : 0;
-        }
-    }
-
-    return 0;
-}
-
-/*---------------------------------------------------------------------------*/
-
-/*
-static void i_dump_area_order(AREA *area, const char_t *label)
-{
-    DBORDERINFO info;
-    memset(&info, 0, sizeof(info));
-    info.itmResult = hb_itemNew(NULL);
-
-    SELF_ORDINFO(area, DBOI_NUMBER, &info);
-    int nOrder = hb_itemGetNI(info.itmResult);
-
-    SELF_ORDINFO(area, DBOI_NAME, &info);
-    const char *szName = hb_itemGetCPtr(info.itmResult);
-    bstd_printf("%s order=%d  tag='%s'\n", label, nOrder, szName);
-    hb_itemRelease(info.itmResult);
-}
-*/
-
-/*---------------------------------------------------------------------------*/
-
-/*
-static void i_dump_area_tags(AREAP area, const char *label)
-{
-    DBORDERINFO info;
-    int i, nTotal;
-
-    memset(&info, 0, sizeof(info));
-    info.itmResult = hb_itemNew(NULL);
-    info.itmOrder  = hb_itemNew(NULL);
-
-    SELF_ORDINFO(area, DBOI_ORDERCOUNT, &info);
-    nTotal = hb_itemGetNI(info.itmResult);
-    bstd_printf("%s: %d tag(s)\n", label, nTotal);
-
-    for (i = 1; i <= nTotal; i++)
-    {
-        hb_itemPutNI(info.itmOrder, i);
-        SELF_ORDINFO(area, DBOI_NAME, &info);
-        bstd_printf("  [%d] '%s'\n", i, hb_itemGetCPtr(info.itmResult));
-    }
-
-    hb_itemRelease(info.itmResult);
-    hb_itemRelease(info.itmOrder);
-}
-*/
-
-/*---------------------------------------------------------------------------*/
-
-static void i_dbconn_refresh(GtNapFDBConn *dbconn)
-{
-    uint32_t nareas = 0;
-    NodeSt(GtNapFNode) *root = NULL;
-    GtNapFNode *vdata = NULL;
-    GtNapFArea2 *area = NULL;
-    HB_BOOL eof = HB_FALSE;
-
-    cassert_no_null(dbconn);
-    treest_clear(dbconn->tdata, i_remove_fnode, GtNapFNode);
-    nareas = arrst_size(dbconn->areas, GtNapFArea2);
-    cassert(nareas > 0);
-
-    /* Virtual root: invisible container for all top-level nodes */
-    root = treest_root_new(dbconn->tdata, GtNapFNode);
-    vdata = treest_node_data(root, GtNapFNode);
-    vdata->area = NULL;
-    vdata->recno = UINT32_MAX;
-    vdata->expanded = TRUE;
-
-    /* Navigate root area and build one top-level node per record */
-    area = arrst_get(dbconn->areas, 0, GtNapFArea2);
-    SELF_GOTOP(area->area);
-    SELF_EOF(area->area, &eof);
-
-    while (eof == HB_FALSE)
-    {
-        HB_ULONG recno = 0;
-        NodeSt(GtNapFNode) *node = NULL;
-        GtNapFNode *data = NULL;
-
-        SELF_RECNO(area->area, &recno);
-        node = treest_node_insert(root, UINT32_MAX, GtNapFNode);
-        data = treest_node_data(node, GtNapFNode);
-        data->area = area;
-        data->recno = recno;
-        if (area->fexpanded != 0)
-        {
-            PHB_ITEM pExp = hb_itemNew(NULL);
-            SELF_GETVALUE(area->area, area->fexpanded, pExp);
-            data->expanded = (bool_t)hb_itemGetL(pExp);
-            hb_itemRelease(pExp);
-        }
-        else
-        {
-            data->expanded = FALSE;
-        }
-
-        /* Exists node hierarchy */
-        data->nchildren = 0;
-        if (nareas > 1)
-        {
-            /* Only recurse into children if this node is expanded */
-            if (data->expanded == TRUE)
-            {
-                i_build_children(dbconn, node, 0);
-            }
-            else
-            {
-                GtNapFArea2 *carea = arrst_get(dbconn->areas, 1, GtNapFArea2);
-                data->nchildren = i_count_db_children(area, recno, carea);
-            }
-        }
-
-        /* Restore root cursor after recursion, then advance */
-        SELF_GOTO(area->area, recno);
-        SELF_SKIP(area->area, 1);
-        SELF_EOF(area->area, &eof);
-    }
-
-    /*i_dump_node_count(dbconn->tdata);*/
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_map_dbconn_to_form(GtNapFDBConn *dbconn)
-{
-    cassert_no_null(dbconn);
-    cassert(dbconn->table == NULL);
-    cassert_no_null(dbconn->form);
-    dbconn->table = nform_get_tableview(dbconn->form->form, tc(dbconn->cellname));
-    if (dbconn->table != NULL)
-    {
-        tableview_tree(dbconn->table, 0);
-        tableview_OnData(dbconn->table, listener(dbconn, i_OnTreeFAreaData, GtNapFDBConn));
-        i_dbconn_refresh(dbconn);
-        tableview_update(dbconn->table);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_tree_bind(GtNapForm *form, const char_t *cell, HB_ITEM *areas, HB_ITEM *relations, HB_ITEM *columns)
-{
-    HB_SIZE nA = UINT32_MAX;
-    HB_SIZE nR = UINT32_MAX;
-    HB_SIZE nB = UINT32_MAX;
-    HB_SIZE i = 0;
-    cassert_no_null(form);
-    cassert(form->area == NULL);
-    cassert(form->dbconn == NULL);
-    cassert(HB_ITEM_TYPE(areas) == HB_IT_ARRAY);
-    cassert(HB_ITEM_TYPE(relations) == HB_IT_ARRAY);
-    cassert(HB_ITEM_TYPE(columns) == HB_IT_ARRAY);
-    form->dbconn = i_create_dbconn(form, cell);
-    nA = hb_arrayLen(areas);
-    nR = hb_arrayLen(relations);
-    nB = hb_arrayLen(columns);
-    cassert(nA > 1);
-    cassert_unref(nR == nA - 1, nR);
-    cassert_unref(nB == nA, nB);
-
-    for (i = 1; i <= nA; ++i)
-    {
-        GtNapFArea2 *area = arrst_new0(form->dbconn->areas, GtNapFArea2);
-
-        /* Database area */
-        {
-            PHB_ITEM area_item = hb_arrayGetItemPtr(areas, i);
-            const char *area_id = NULL;
-            int iArea = 0;
-            HB_ERRCODE hbres;
-            cassert(HB_ITEM_TYPE(area_item) == HB_IT_STRING);
-            area_id = hb_itemGetCPtr(area_item);
-            hbres = hb_rddGetAliasNumber(area_id, &iArea);
-            cassert_unref(hbres == HB_SUCCESS, hbres);
-            area->area = cast(hb_rddGetWorkAreaPointer(iArea), AREA);
-            cassert_no_null(area->area);
-            area->fexpanded = i_area_expanded_field(area->area);
-        }
-
-        /* Area->Columns data mappings */
-        {
-            PHB_ITEM bind_item = hb_arrayGetItemPtr(columns, i);
-            HB_SIZE j, nCols = UINT32_MAX;
-            area->columns = arrst_create(GtNapFColumn);
-            cassert(HB_ITEM_TYPE(bind_item) == HB_IT_ARRAY);
-            nCols = hb_arrayLen(bind_item);
-            for (j = 1; j <= nCols; ++j)
-            {
-                GtNapFColumn *col = arrst_new0(area->columns, GtNapFColumn);
-                PHB_ITEM col_item = hb_arrayGetItemPtr(bind_item, j);
-                PHB_ITEM align_item = NULL;
-                PHB_ITEM block_item = NULL;
-                cassert(HB_ITEM_TYPE(col_item) == HB_IT_ARRAY);
-                cassert(hb_arrayLen(col_item) == 2);
-                align_item = hb_arrayGetItemPtr(col_item, 1);
-                block_item = hb_arrayGetItemPtr(col_item, 2);
-                cassert(HB_ITEM_TYPE(align_item) == HB_IT_INTEGER);
-                cassert(HB_ITEM_TYPE(block_item) == HB_IT_BLOCK);
-                col->align = i_hbalign(hb_itemGetNI(align_item));
-                col->block = block_item ? hb_itemNew(block_item) : NULL;
-            }
-        }
-
-        /* Relation with next area: { parent_key_block, "cdx_path", "CHILD_TAG" } */
-        if (i < nA)
-        {
-            PHB_ITEM rel_item = hb_arrayGetItemPtr(relations, i);
-            PHB_ITEM key_item = NULL;
-            PHB_ITEM cdx_item = NULL;
-            PHB_ITEM tag_item = NULL;
-            const char *cdx_path = NULL;
-            const char *tag_name = NULL;
-            cassert(HB_ITEM_TYPE(rel_item) == HB_IT_ARRAY);
-            cassert(hb_arrayLen(rel_item) == 3);
-            key_item = hb_arrayGetItemPtr(rel_item, 1);
-            cdx_item = hb_arrayGetItemPtr(rel_item, 2);
-            tag_item = hb_arrayGetItemPtr(rel_item, 3);
-            cassert(HB_ITEM_TYPE(key_item) == HB_IT_BLOCK);
-            cassert(HB_ITEM_TYPE(cdx_item) == HB_IT_STRING);
-            cassert(HB_ITEM_TYPE(tag_item) == HB_IT_STRING);
-            area->relkey = hb_itemNew(key_item);
-            cdx_path = hb_itemGetCPtr(cdx_item);
-            tag_name = hb_itemGetCPtr(tag_item);
-
-            /* Open CDX and activate tag on the child area */
-            {
-                HB_ERRCODE hbres;
-                int iCArea = 0;
-                AREA *carea = NULL;
-
-                {
-                    PHB_ITEM carea_item = hb_arrayGetItemPtr(areas, i + 1);
-                    cassert(HB_ITEM_TYPE(carea_item) == HB_IT_STRING);
-                    hbres = hb_rddGetAliasNumber(hb_itemGetCPtr(carea_item), &iCArea);
-                    cassert_unref(hbres == HB_SUCCESS, hbres);
-                    carea = cast(hb_rddGetWorkAreaPointer(iCArea), AREA);
-                    cassert_no_null(carea);
-                }
-
-                /*
-                 * Open CDX and activate the specified tag.
-                 * SELF_ORDLSTADD evaluates the key expression via the Harbour VM
-                 * macro compiler, which resolves fields against the CURRENTLY
-                 * SELECTED work area — not the area pointer argument. We must
-                 * select the child area first so field names resolve correctly.
-                 */
-                {
-                    DBORDERINFO oinfo;
-                    int iSavedArea = hb_rddGetCurrentWorkAreaNumber();
-                    hbres = hb_rddSelectWorkAreaNumber(iCArea);
-                    cassert_unref(hbres == HB_SUCCESS, hbres);
-                    hbres = SELF_ORDLSTCLEAR(carea);
-                    cassert_unref(hbres == HB_SUCCESS, hbres);
-                    memset(&oinfo, 0, sizeof(oinfo));
-                    oinfo.atomBagName = hb_itemPutC(NULL, cdx_path);
-                    oinfo.itmResult = hb_itemNew(NULL);
-                    hbres = SELF_ORDLSTADD(carea, &oinfo);
-                    cassert_unref(hbres == HB_SUCCESS, hbres);
-                    hb_itemRelease(oinfo.atomBagName);
-                    oinfo.atomBagName = NULL;
-                    oinfo.itmOrder = hb_itemPutC(NULL, tag_name);
-                    hbres = SELF_ORDLSTFOCUS(carea, &oinfo);
-                    cassert_unref(hbres == HB_SUCCESS, hbres);
-                    hb_itemRelease(oinfo.itmOrder);
-                    hb_itemRelease(oinfo.itmResult);
-
-                    /*
-                     * ORDLSTFOCUS returns HB_SUCCESS even if tag name does not exist
-                     * verify by checking the active order number
-                     */
-                    {
-                        DBORDERINFO vinfo;
-                        int pos = 0;
-                        memset(&vinfo, 0, sizeof(vinfo));
-                        vinfo.itmResult = hb_itemNew(NULL);
-                        hbres = SELF_ORDINFO(carea, DBOI_NUMBER, &vinfo);
-                        cassert_unref(hbres == HB_SUCCESS, hbres);
-                        pos = hb_itemGetNI(vinfo.itmResult);
-                        cassert_unref(pos != 0, pos);
-                        hb_itemRelease(vinfo.itmResult);
-                    }
-
-                    hbres = hb_rddSelectWorkAreaNumber(iSavedArea);
-                    cassert_unref(hbres == HB_SUCCESS, hbres);
-                }
-            }
-        }
-        else
-        {
-            area->relkey = NULL;
-        }
-    }
-
-    i_map_dbconn_to_form(form->dbconn);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_item_list(GtNapForm *form, const char_t *cell, HB_ITEM *items)
-{
-    HB_SIZE i, n = UINT32_MAX;
-    cassert_no_null(form);
-    cassert(HB_ITEM_TYPE(items) == HB_IT_ARRAY);
-    nform_clear_control_list(form->form, cell);
-    n = hb_arrayLen(items);
-    for (i = 1; i <= n; ++i)
-    {
-        PHB_ITEM hitem = hb_arrayGetItemPtr(items, i);
-        const char_t *text = NULL;
-        cassert(HB_ITEM_TYPE(hitem) == HB_IT_STRING);
-        text = hb_itemGetCPtr(hitem);
-        nform_add_control_item(form->form, cell, text);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static Listener *i_gtnap_form_listener(HB_ITEM *block, GtNapForm *form, FPtr_gtnap_callback func_callback)
-{
-    GtNapCallback *callback = heap_new0(GtNapCallback);
-    cassert_no_null(form);
-    callback->block = block ? hb_itemNew(block) : NULL;
-    callback->form = form;
-    callback->key = INT32_MAX;
-    callback->autoclose_id = UINT32_MAX;
-    arrpt_append(form->callbacks, callback, GtNapCallback);
-    return listener(callback, func_callback, GtNapCallback);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnFormButtonClick(GtNapCallback *callback, Event *e)
-{
-    cassert_no_null(callback);
-    cassert_no_null(callback->form);
-    unref(e);
-    if (callback->block != NULL)
-    {
-        PHB_ITEM ritem = hb_itemDo(callback->block, 0);
-        hb_itemRelease(ritem);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_OnClick(GtNapForm *form, const char_t *cell, HB_ITEM *click_block)
-{
-    Listener *listener = i_gtnap_form_listener(click_block, form, i_OnFormButtonClick);
-    cassert_no_null(form);
-    nform_set_listener(form->form, cell, listener);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_maximize(GtNapForm *form)
-{
-    cassert_no_null(form);
-    window_maximize(form->window);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnFormClose(GtNapForm *form, Event *e)
-{
-    cassert_no_null(form);
-    if (form->OnClose_block != NULL)
-    {
-        PHB_ITEM ritem = hb_itemDo(form->OnClose_block, 0);
-        if (HB_ITEM_TYPE(ritem) == HB_IT_LOGICAL)
-        {
-            bool_t *r = event_result(e, bool_t);
-            *r = (bool_t)hb_itemGetL(ritem);
-        }
-        hb_itemRelease(ritem);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static V2Df i_center_window(const Window *parent, Window *window)
-{
-    V2Df p1 = window_get_origin(parent);
-    S2Df s1 = window_get_size(parent);
-    S2Df s2 = window_get_size(window);
-    V2Df p2;
-    p2.x = p1.x + (s1.width - s2.width) / 2;
-    p2.y = p1.y + (s1.height - s2.height) / 2;
-    return p2;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_form_frame(GtNapForm *form, Window *parent)
-{
-    S2Df size;
-    cassert_no_null(form);
-    size.width = i_read_prop_r32(tc(form->nameid), i_WIDTH_PROP);
-    size.height = i_read_prop_r32(tc(form->nameid), i_HEIGHT_PROP);
-    if ((size.width == i_MAXIMIZED_SIZE || size.height == i_MAXIMIZED_SIZE) && form->is_resizable == TRUE)
-    {
-        window_maximize(form->window);
-    }
-    else
-    {
-        V2Df pos;
-        pos.x = i_read_prop_r32(tc(form->nameid), i_XPOS_PROP);
-        pos.y = i_read_prop_r32(tc(form->nameid), i_YPOS_PROP);
-
-        if ((pos.x == i_UNDEF_R32 || pos.y == i_UNDEF_R32) && parent != NULL)
-            pos = i_center_window(parent, form->window);
-
-        if ((size.width != i_UNDEF_R32 && size.height != i_UNDEF_R32) && form->is_resizable == TRUE)
-            window_client_size(form->window, size);
-
-        window_origin(form->window, pos);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_show(GtNapForm *form, HB_ITEM *onclose_block)
-{
-    cassert_no_null(form);
-    if (form->OnClose_block != NULL)
-        hb_itemRelease(form->OnClose_block);
-
-    form->OnClose_block = hb_itemNew(onclose_block);
-    window_OnClose(form->window, listener(form, i_OnFormClose, GtNapForm));
-    window_update(form->window);
-    i_form_frame(form, NULL);
-    window_show(form->window);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static uint32_t i_forms_modal(GtNapForm *form, Window *parent)
-{
-    cassert_no_null(form);
-    window_update(form->window);
-    i_form_frame(form, parent);
-    form->modal_ret = window_modal(form->window, parent);
-    return form->modal_ret;
-}
-
-/*---------------------------------------------------------------------------*/
-
-uint32_t hbnap_forms_modal(GtNapForm *form, GtNapForm *parent)
-{
-    cassert_no_null(parent);
-    return i_forms_modal(form, parent->window);
-}
-
-/*---------------------------------------------------------------------------*/
-
-uint32_t hbnap_forms_modal_gtnap(GtNapForm *form)
-{
-    GtNapWindow *gtwin = i_current_gtwin(GTNAP_GLOBAL);
-    cassert_no_null(gtwin);
-    return i_forms_modal(form, gtwin->window);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_stop_modal(GtNapForm *form, const uint32_t value)
-{
-    cassert_no_null(form);
-    window_stop_modal(form->window, value);
-}
-
-/*---------------------------------------------------------------------------*/
-
-R2Df hbnap_forms_control_frame(GtNapForm *form, const char_t *cell)
-{
-    cassert_no_null(form);
-    return nform_get_control_frame(form->form, cell, form->window);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_update(GtNapForm *form)
-{
-    cassert_no_null(form);
-    window_update(form->window);
-}
-
-/*---------------------------------------------------------------------------*/
-
-typedef struct i_mainitem_t MainItem;
-typedef struct i_maindata_t MainData;
-
-struct i_mainitem_t
-{
-    String *title;
-    String *more;
-    color_t back;
-    Image *icon;
-    bool_t isnew;
-    HB_ITEM *OnClick;
-    real32_t pos_x;
-    real32_t pos_y;
-    real32_t icon_scale;
-};
-
-struct i_maindata_t
-{
-    String *title;
-    Image *logo;
-    real32_t title_width;
-    real32_t title_height;
-    real32_t logo_width;
-    real32_t logo_height;
-    Font *title_font;
-    Font *item_font;
-    Font *item_sfont;
-    uint32_t hover_item;
-    ArrSt(MainItem) *items;
-};
-
-/*---------------------------------------------------------------------------*/
-
-DeclSt(MainItem);
-static const real32_t i_ITEM_WIDTH = 260;
-static const real32_t i_ITEM_HEIGHT = 160;
-static const real32_t i_ITEM_SEP = 16;
-static const real32_t i_ITEM_CIRCLE_RADIX = 50;
-static const real32_t i_ITEM_CIRCLE_OFFSET = 14;
-static const real32_t i_ITEM_ICON_SIZE = 64;
-static const real32_t i_ITEM_TEXT_PADDING_BOTTOM = 12;
-static const real32_t i_IMAGE_PADDING_TOP = 30;
-static const real32_t i_IMAGE_PADDING_BOTTOM = 36;
-static const real32_t i_IMAGE_PADDING_RIGHT = 70;
-static const real32_t i_TITLE_PADDING_LEFT = 80;
-
-/*---------------------------------------------------------------------------*/
-
-static void i_remove_mainitem(MainItem *item)
-{
-    str_destroy(&item->title);
-    str_destroy(&item->more);
-    ptr_destopt(image_destroy, &item->icon, Image);
-    if (item->OnClick != NULL)
-    {
-        hb_itemRelease(item->OnClick);
-        item->OnClick = NULL;
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_destroy_maindata(MainData **data)
-{
-    cassert_no_null(data);
-    cassert_no_null(*data);
-    str_destopt(&(*data)->title);
-    font_destroy(&(*data)->title_font);
-    font_destroy(&(*data)->item_font);
-    font_destroy(&(*data)->item_sfont);
-    ptr_destopt(image_destroy, &(*data)->logo, Image);
-    arrst_destroy(&(*data)->items, i_remove_mainitem, MainItem);
-    heap_delete(data, MainData);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_draw_mainitem(DCtx *ctx, const MainItem *item, const Font *font, const Font *sfont, const T2Df *t2d)
-{
-    color_t color = kCOLOR_WHITE;
-    cassert_no_null(item);
-    draw_matrixf(ctx, t2d);
-    draw_fill_color(ctx, item->back);
-    draw_line_color(ctx, color);
-    draw_text_color(ctx, color);
-    draw_line_width(ctx, 3);
-    draw_rect(ctx, ekFILL, 0, 0, i_ITEM_WIDTH, i_ITEM_HEIGHT);
-    draw_circle(ctx, ekSTROKE, i_ITEM_WIDTH / 2, (i_ITEM_HEIGHT / 2) - i_ITEM_CIRCLE_OFFSET, i_ITEM_CIRCLE_RADIX);
-
-    if (item->icon != NULL)
-    {
-        T2Df scale;
-        t2d_movef(&scale, t2d, (i_ITEM_WIDTH - i_ITEM_ICON_SIZE) / 2, (i_ITEM_HEIGHT - i_ITEM_ICON_SIZE) / 2 - i_ITEM_CIRCLE_OFFSET);
-        t2d_scalef(&scale, &scale, item->icon_scale, item->icon_scale);
-        draw_matrixf(ctx, &scale);
-        draw_image(ctx, item->icon, 0, 0);
-        draw_matrixf(ctx, t2d);
-    }
-
-    if (str_empty(item->title) == FALSE)
-    {
-        draw_font(ctx, font);
-        draw_text_align(ctx, ekCENTER, ekBOTTOM);
-        draw_text(ctx, tc(item->title), i_ITEM_WIDTH / 2, i_ITEM_HEIGHT - i_ITEM_TEXT_PADDING_BOTTOM);
-    }
-
-    if (str_empty(item->more) == FALSE)
-    {
-        draw_font(ctx, sfont);
-        draw_text_align(ctx, ekRIGHT, ekTOP);
-        draw_text(ctx, tc(item->more), i_ITEM_WIDTH, 0);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static T2Df i_item_transform(const MainItem *item, const bool_t hover)
-{
-    T2Df t2d;
-    t2d_movef(&t2d, kT2D_IDENTf, item->pos_x, item->pos_y);
-
-    if (hover == TRUE)
-    {
-        t2d_movef(&t2d, &t2d, i_ITEM_WIDTH / 2, i_ITEM_HEIGHT / 2);
-        t2d_scalef(&t2d, &t2d, 1.05f, 1.05f);
-        t2d_movef(&t2d, &t2d, - i_ITEM_WIDTH / 2, - i_ITEM_HEIGHT / 2);
-    }
-
-    return t2d;
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnDrawMainView(GtNapForm *form, Event *e)
-{
-    const EvDraw *p = event_params(e, EvDraw);
-    View *view = event_sender(e, View);
-    MainData *data = view_get_data(view, MainData);
-    real32_t logo_x = 0;
-    bool_t draw_title = FALSE;
-    color_t text_color;
-    cassert_no_null(data);
-    unref(form);
-
-    if (gui_dark_mode() == TRUE)
-        text_color = color_rgb(255, 255, 255);
-    else
-        text_color = color_rgb(99, 99, 99);
-
-    logo_x = p->width - data->logo_width - i_IMAGE_PADDING_RIGHT;
-
-    if (str_empty(data->title) == FALSE)
-    {
-        draw_text_color(p->ctx, text_color);
-        draw_line_color(p->ctx, text_color);
-        draw_font(p->ctx, data->title_font);
-
-        if (data->title_width == 0)
-            draw_text_extents(p->ctx, tc(data->title), -1, &data->title_width, &data->title_height);
-
-        /* Avoid title and logo overlap */
-        if (data->title_width + i_TITLE_PADDING_LEFT < logo_x)
-            draw_title = TRUE;
-
-        if (draw_title == TRUE)
-        {
-            real32_t line_y = i_IMAGE_PADDING_TOP + data->logo_height;
-            draw_text(p->ctx, tc(data->title), i_TITLE_PADDING_LEFT, i_IMAGE_PADDING_TOP + data->logo_height - data->title_height);
-            draw_line(p->ctx, i_TITLE_PADDING_LEFT, line_y, i_TITLE_PADDING_LEFT + data->title_width + data->logo_width, line_y);
-        }
-    }
-
-    if (data->logo != NULL)
-    {
-        /* If no title is drawn, the logo is centered */
-        if (draw_title == FALSE)
-            logo_x = (p->width - data->logo_width) / 2;
-        draw_image(p->ctx, data->logo, logo_x, i_IMAGE_PADDING_TOP);
-    }
-
-    arrst_foreach_const(item, data->items, MainItem)
-        T2Df t2d = i_item_transform(item, (bool_t)(item_i == data->hover_item));
-        i_draw_mainitem(p->ctx, item, data->item_font, data->item_sfont, &t2d);
-    arrst_end()
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_mainitems_locations(View *view, const real32_t width)
-{
-    MainData *data = view_get_data(view, MainData);
-    real32_t content_width = 0;
-    real32_t content_height = 0;
-    uint32_t n = 0, ncols, nrows;
-    cassert_no_null(data);
-    n = arrst_size(data->items, MainItem);
-
-    /* Compute the number of columns */
-    for (ncols = 4; ncols >= 1; --ncols)
-    {
-        content_width = i_ITEM_WIDTH * ncols + i_ITEM_SEP * (ncols - 1);
-        if (content_width < width || ncols == 1)
-            break;
-    }
-
-    cassert(ncols > 0);
-    nrows = (n / ncols) + ((n % ncols) ? 1 : 0);
-    content_height = i_IMAGE_PADDING_TOP + i_ITEM_HEIGHT * nrows + i_ITEM_SEP * (nrows - 1);
-    content_height += data->logo_height;
-
-    if (data->logo_height > 0)
-        content_height += i_IMAGE_PADDING_BOTTOM;
-
-    content_height += i_IMAGE_PADDING_BOTTOM;
-
-    /* Compute the items position */
-    arrst_foreach(item, data->items, MainItem)
-        uint32_t i = item_i % ncols;
-        uint32_t j = item_i / ncols;
-        item->pos_x = ((width - content_width) / 2) + i * (i_ITEM_WIDTH + i_ITEM_SEP);
-        item->pos_y = i_IMAGE_PADDING_TOP + data->logo_height + j * (i_ITEM_HEIGHT + i_ITEM_SEP);
-        if (data->logo_height > 0)
-            item->pos_y += i_IMAGE_PADDING_BOTTOM;
-    arrst_end()
-
-    view_content_size(view, s2df(content_width, content_height), s2df(10, 10));
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnSizeMainView(GtNapForm *form, Event *e)
-{
-    const EvSize *p = event_params(e, EvSize);
-    View *view = event_sender(e, View);
-    i_mainitems_locations(view, p->width);
-    unref(form);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnMoveMainView(GtNapForm *form, Event *e)
-{
-    const EvMouse *p = event_params(e, EvMouse);
-    View *view = event_sender(e, View);
-    MainData *data = view_get_data(view, MainData);
-    cassert_no_null(data);
-    data->hover_item = UINT32_MAX;
-    arrst_foreach_const(item, data->items, MainItem)
-        R2Df r2d = r2df(item->pos_x, item->pos_y, i_ITEM_WIDTH, i_ITEM_HEIGHT);
-        if (r2d_containsf(&r2d, p->x, p->y) == TRUE)
-        {
-            data->hover_item = item_i;
-            break;
-        }
-    arrst_end()
-    view_update(view);
-    unref(form);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnClickMainView(GtNapForm *form, Event *e)
-{
-    View *view = event_sender(e, View);
-    MainData *data = view_get_data(view, MainData);
-    cassert_no_null(data);
-    if (data->hover_item != UINT32_MAX)
-    {
-        const MainItem *item = arrst_get_const(data->items, data->hover_item, MainItem);
-        PHB_ITEM ritem = hb_itemDo(item->OnClick, 0);
-        hb_itemRelease(ritem);
-    }
-    unref(form);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_forms_main_cover(GtNapForm *form, const char_t *canvas_cell, const char_t *title, const char_t *logo_path, HB_ITEM *cover_items)
-{
-    View *view = NULL;
-    S2Df view_size;
-    MainData *data = heap_new0(MainData);
-    HB_SIZE i, n;
-    cassert_no_null(form);
-    view = nform_get_view(form->form, canvas_cell);
-    cassert_no_null(view);
-    cassert_no_null(cover_items);
-    cassert(HB_ITEM_TYPE(cover_items) == HB_IT_ARRAY);
-    data->title = str_c(title);
-    data->title_font = font_system(42, 0);
-    data->item_font = font_system(20, ekFBOLD);
-    data->item_sfont = font_system(14, ekFBOLD);
-    data->logo = image_from_file(logo_path, NULL);
-
-    if (data->logo != NULL)
-    {
-        data->logo_width = (real32_t)image_width(data->logo);
-        data->logo_height = (real32_t)image_height(data->logo);
-    }
-
-    data->items = arrst_create(MainItem);
-    data->hover_item = UINT32_MAX;
-
-    n = hb_arrayLen(cover_items);
-    for (i = 1; i <= n; ++i)
-    {
-        MainItem *item = arrst_new0(data->items, MainItem);
-        HB_ITEM *citem = hb_arrayGetItemPtr(cover_items, i);
-        HB_ITEM *ititle = NULL;
-        HB_ITEM *icon = NULL;
-        HB_ITEM *color = NULL;
-        HB_ITEM *more = NULL;
-        HB_ITEM *isnew = NULL;
-        HB_ITEM *block = NULL;
-        cassert_no_null(citem);
-        cassert(HB_ITEM_TYPE(citem) == HB_IT_ARRAY);
-        cassert(hb_arrayLen(citem) == 6);
-        ititle = hb_arrayGetItemPtr(citem, 1);
-        icon = hb_arrayGetItemPtr(citem, 2);
-        color = hb_arrayGetItemPtr(citem, 3);
-        more = hb_arrayGetItemPtr(citem, 4);
-        isnew = hb_arrayGetItemPtr(citem, 5);
-        block = hb_arrayGetItemPtr(citem, 6);
-        cassert(HB_ITEM_TYPE(ititle) == HB_IT_STRING);
-        cassert(HB_ITEM_TYPE(icon) == HB_IT_STRING);
-        cassert(HB_ITEM_TYPE(color) == HB_IT_STRING);
-        cassert(HB_ITEM_TYPE(more) == HB_IT_STRING);
-        cassert(HB_ITEM_TYPE(isnew) == HB_IT_LOGICAL);
-        cassert(HB_ITEM_TYPE(block) == HB_IT_BLOCK);
-        item->title = str_c(hb_itemGetCPtr(ititle));
-        item->icon = image_from_file(hb_itemGetCPtr(icon), NULL);
-        item->back = color_html(hb_itemGetCPtr(color));
-        item->more = str_c(hb_itemGetCPtr(more));
-        item->isnew = (bool_t)hb_itemGetL(isnew);
-        item->OnClick = hb_itemNew(block);
-
-        if (item->icon != NULL)
-        {
-            real32_t width = (real32_t)image_width(item->icon);
-            item->icon_scale = i_ITEM_ICON_SIZE / width;
-        }
-    }
-
-    view_data(view, &data, i_destroy_maindata, MainData);
-    view_OnDraw(view, listener(form, i_OnDrawMainView, GtNapForm));
-    view_OnSize(view, listener(form, i_OnSizeMainView, GtNapForm));
-    view_OnMove(view, listener(form, i_OnMoveMainView, GtNapForm));
-    view_OnClick(view, listener(form, i_OnClickMainView, GtNapForm));
-    view_get_size(view, &view_size);
-    i_mainitems_locations(view, view_size.width);
-}
-
-/*---------------------------------------------------------------------------*/
-
-GtNapMenu *hbnap_menu_create(void)
-{
-    Menu *menu = menu_create();
-    return cast(menu, GtNapMenu);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_remove_menu_callbacks(Menu *menu);
-
-/*---------------------------------------------------------------------------*/
-
-static void i_remove_item_callbacks(MenuItem *item)
-{
-    Menu *submenu = menuitem_get_submenu(item);
-    uint32_t callback_id = UINT32_MAX;
-
-    arrpt_foreach(callback, GTNAP_GLOBAL->menu_callbacks, GtNapCallback)
-        if (callback->menuitem == cast(item, GtNapMenuItem))
-        {
-            callback_id = callback_i;
-            break;
-        }
-    arrpt_end()
-
-    if (callback_id != UINT32_MAX)
-        arrpt_delete(GTNAP_GLOBAL->menu_callbacks, callback_id, i_destroy_callback, GtNapCallback);
-
-    if (submenu != NULL)
-        i_remove_menu_callbacks(submenu);
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_remove_menu_callbacks(Menu *menu)
-{
-    uint32_t i, n = menu_count(menu);
-    for (i = 0; i < n; ++i)
-    {
-        MenuItem *item = menu_get_item(menu, i);
-        i_remove_item_callbacks(item);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_menu_destroy(GtNapMenu *menu)
-{
-    i_remove_menu_callbacks(cast(menu, Menu));
-    menu_destroy(dcast(&menu, Menu));
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_menu_add_item(GtNapMenu *menu, GtNapMenuItem *item)
-{
-    menu_add_item(cast(menu, Menu), cast(item, MenuItem));
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_menu_ins_item(GtNapMenu *menu, const uint32_t pos, GtNapMenuItem *item)
-{
-    menu_ins_item(cast(menu, Menu), pos, cast(item, MenuItem));
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_menu_del_item(GtNapMenu *menu, const uint32_t pos)
-{
-    MenuItem *item = menu_get_item(cast(menu, Menu), pos);
-    i_remove_item_callbacks(item);
-    menu_del_item(cast(menu, Menu), pos);
-}
-
-/*---------------------------------------------------------------------------*/
-
-uint32_t hbnap_menu_count(const GtNapMenu *menu)
-{
-    return menu_count(cast(menu, Menu));
-}
-
-/*---------------------------------------------------------------------------*/
-
-GtNapMenuItem *hbnap_menu_get_item(GtNapMenu *menu, const uint32_t index)
-{
-    MenuItem *item = menu_get_item(cast(menu, Menu), index);
-    return cast(item, GtNapMenuItem);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_menubar(GtNapMenu *menu, GtNapForm *form)
-{
-    cassert_no_null(form);
-    osapp_menubar(cast(menu, Menu), form->window);
-}
-
-/*---------------------------------------------------------------------------*/
-
-bool_t hbnap_is_menubar(const GtNapMenu *menu)
-{
-    return menu_is_menubar(cast_const(menu, Menu));
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_menu_popup(GtNapMenu *menu, GtNapForm *form, const int32_t x, const int32_t y)
-{
-    cassert_no_null(form);
-    menu_launch(cast(menu, Menu), form->window, v2df((real32_t)x, (real32_t)y));
-}
-
-/*---------------------------------------------------------------------------*/
-
-static void i_OnMenuClick(GtNapCallback *callback, Event *e)
-{
-    cassert_no_null(callback);
-    unref(e);
-    if (callback->block != NULL)
-    {
-        /* The menuitem itself will be allways the first param in click callback */
-        PHB_ITEM pItem = hb_itemPutPtr(NULL, callback->menuitem);
-        PHB_ITEM ritem = hb_itemDo(callback->block, 1, pItem);
-        hb_itemRelease(pItem);
-        hb_itemRelease(ritem);
-    }
-}
-
-/*---------------------------------------------------------------------------*/
-
-static Listener *i_gtnap_menu_listener(HB_ITEM *block, GtNapMenuItem *item)
-{
-    GtNapCallback *callback = heap_new0(GtNapCallback);
-    cassert_no_null(item);
-    callback->block = block ? hb_itemNew(block) : NULL;
-    callback->menuitem = item;
-    callback->key = INT32_MAX;
-    callback->autoclose_id = UINT32_MAX;
-    arrpt_append(GTNAP_GLOBAL->menu_callbacks, callback, GtNapCallback);
-    return listener(callback, i_OnMenuClick, GtNapCallback);
-}
-
-/*---------------------------------------------------------------------------*/
-
-GtNapMenuItem *hbnap_menuitem_create(const char_t *text, const char_t *icon_path, HB_ITEM *click_block)
-{
-    MenuItem *item = menuitem_create();
-    menuitem_text(item, text);
-
-    if (str_empty_c(icon_path) == FALSE)
-    {
-        Image *image = image_from_file(icon_path, NULL);
-        if (image != NULL)
-        {
-            menuitem_image(item, image);
-            image_destroy(&image);
-        }
-    }
-
-    {
-        Listener *listener = i_gtnap_menu_listener(click_block, cast(item, GtNapMenuItem));
-        menuitem_OnClick(item, listener);
-    }
-
-    return cast(item, GtNapMenuItem);
-}
-
-/*---------------------------------------------------------------------------*/
-
-GtNapMenuItem *hbnap_menuitem_separator(void)
-{
-    MenuItem *item = menuitem_separator();
-    return cast(item, GtNapMenuItem);
-}
-
-/*---------------------------------------------------------------------------*/
-
-void hbnap_menuitem_submenu(GtNapMenuItem *item, GtNapMenu *submenu)
-{
-    menuitem_submenu(cast(item, MenuItem), dcast(&submenu, Menu));
-}
-
-/*---------------------------------------------------------------------------*/
-
-const char_t *hbnap_menuitem_get_text(const GtNapMenuItem *item)
-{
-    return menuitem_get_text(cast_const(item, MenuItem));
-}
-
-/*---------------------------------------------------------------------------*/
-
-GtNapMenu *hbnap_menuitem_get_submenu(GtNapMenuItem *item)
-{
-    Menu *menu = menuitem_get_submenu(cast(item, MenuItem));
-    return cast(menu, GtNapMenu);
 }
 
 /*---------------------------------------------------------------------------*/
